Refactor: FormatDate 파일 분리
